--- a/cmd/dagger/module.go
+++ b/cmd/dagger/module.go
@@ -143,7 +143,7 @@ If --sdk is specified, the given SDK is installed in the module. You can do this
 			dag := engineClient.Dagger()
 
 			// default the module source root to the current working directory if it doesn't exist yet
-			cwd, err := os.Getwd()
+			cwd, err := client.Getwd()
 			if err != nil {
 				return fmt.Errorf("failed to get current working directory: %w", err)
 			}
@@ -191,7 +191,7 @@ If --sdk is specified, the given SDK is installed in the module. You can do this
 
 				if moduleSourcePath != "" {
 					// ensure source path is relative to the source root
-					sourceAbsPath, err := filepath.Abs(moduleSourcePath)
+					sourceAbsPath, err := client.Abs(moduleSourcePath)
 					if err != nil {
 						return fmt.Errorf("failed to get absolute source path for %s: %w", moduleSourcePath, err)
 					}
@@ -263,7 +263,7 @@ var moduleInstallCmd = &cobra.Command{
 			}
 			if depSrcKind == dagger.ModuleSourceKindLocalSource {
 				// need to ensure that local dep paths are relative to the parent root source
-				depAbsPath, err := filepath.Abs(depRefStr)
+				depAbsPath, err := client.Abs(depRefStr)
 				if err != nil {
 					return fmt.Errorf("failed to get dep absolute path for %s: %w", depRefStr, err)
 				}
@@ -459,7 +459,7 @@ This command is idempotent: you can run it at any time, any number of times. It 
 			}
 			if developSourcePath != "" {
 				// ensure source path is relative to the source root
-				sourceAbsPath, err := filepath.Abs(developSourcePath)
+				sourceAbsPath, err := client.Abs(developSourcePath)
 				if err != nil {
 					return fmt.Errorf("failed to get absolute source path for %s: %w", developSourcePath, err)
 				}
@@ -747,12 +747,12 @@ func getModuleConfigurationForSourceRef(
 		srcRefStr = findupConfigDir
 	}
 
-	conf.LocalRootSourcePath, err = filepath.Abs(srcRefStr)
+	conf.LocalRootSourcePath, err = client.Abs(srcRefStr)
 	if err != nil {
 		return nil, fmt.Errorf("failed to get absolute path for %s: %w", srcRefStr, err)
 	}
 	if filepath.IsAbs(srcRefStr) {
-		cwd, err := os.Getwd()
+		cwd, err := client.Getwd()
 		if err != nil {
 			return nil, fmt.Errorf("failed to get current working directory: %w", err)
 		}
@@ -806,7 +806,7 @@ func findUp(curDirPath string) (string, bool, error) {
 	}
 
 	// didn't exist, try parent unless we've hit the root or a git repo checkout root
-	curDirAbsPath, err := filepath.Abs(curDirPath)
+	curDirAbsPath, err := client.Abs(curDirPath)
 	if err != nil {
 		return "", false, fmt.Errorf("failed to get absolute path for %s: %w", curDirPath, err)
 	}