--- a/core/modulesource.go
+++ b/core/modulesource.go
@@ -515,9 +515,11 @@ func callerHostFindUpContext(
 	bk *buildkit.Client,
 	curDirPath string,
 ) (string, bool, error) {
-	_, err := bk.StatCallerHostPath(ctx, filepath.Join(curDirPath, ".git"), false)
+	stat, err := bk.StatCallerHostPath(ctx, filepath.Join(curDirPath, ".git"), true)
 	if err == nil {
-		return curDirPath, true, nil
+		// NOTE: important that we use stat.Path here rather than curDirPath since the stat also
+		// does some normalization of paths when the client is using case-insensitive filesystems
+		return filepath.Dir(stat.Path), true, nil
 	}
 	// TODO: remove the strings.Contains check here (which aren't cross-platform),
 	// since we now set NotFound (since v0.11.2)