--- a/dagql/idtui/testdata/TestTelemetry/TestGolden/viztest/broken/broken
+++ b/dagql/idtui/testdata/TestTelemetry/TestGolden/viztest/broken/broken
@@ -12,6 +12,8 @@ Expected stderr:
 │ ! input: module.withSource.initialize failed to initialize module: failed to call module "broken" to get functions: call constructor: process "go build -ldflags -s -w -o /runtime ." did not complete successfully: exit code: 1
 │ │ $ ModuleSource.resolveFromCaller: ModuleSource! X.Xs CACHED
 │ │ │ ✔ upload /XXX/XXX/XXX from XXXXXXXXXXX (client id: XXXXXXXXXXX, session id: XXXXXXXXXXX) (include: XXXXXXXXXXX) (exclude: XXXXXXXXXXX) X.Xs
+│ │ │ │ ✔ filesync X.Xs
+│ │ │ │ │ ✔ copy X.Xs
 │ │ ✘ .asModule: Module! X.Xs
 │ │ ! process "go build -ldflags -s -w -o /runtime ." did not complete successfully: exit code: 1
 │ │ │ ✘ Module.withSource(