--- a/engine/buildkit/tracing.go
+++ b/engine/buildkit/tracing.go
@@ -1,0 +1,13 @@
+package buildkit
+
+import (
+	"context"
+
+	"go.opentelemetry.io/otel/trace"
+)
+
+const InstrumentationLibrary = "dagger.io/engine.buildkit"
+
+func Tracer(ctx context.Context) trace.Tracer {
+	return trace.SpanFromContext(ctx).TracerProvider().Tracer(InstrumentationLibrary)
+}