--- a/engine/sources/local/fs.go
+++ b/engine/sources/local/fs.go
@@ -1,0 +1,16 @@
+package local
+
+import (
+	"context"
+	"io"
+	"io/fs"
+)
+
+type WalkFS interface {
+	Walk(ctx context.Context, path string, walkFn fs.WalkDirFunc) error
+}
+
+type ReadFS interface {
+	WalkFS
+	ReadFile(ctx context.Context, path string) (io.ReadCloser, error)
+}