… .changes/unreleased/Added-20241126-203335.yaml
… core/integration/local_test.go
… docs/static/api/reference/index.html
… engine/buildkit/containerimage.go
… engine/buildkit/executor_spec.go
… engine/buildkit/filesync.go
… engine/buildkit/llbdefinition.go
… engine/client/filesync.go
… engine/client/pathutil.go
… engine/contenthash/contenthash.go
… engine/opts.go
… engine/server/bk_session.go
… engine/server/server.go
… engine/server/session.go
… engine/sources/blob/blobsource.go
… engine/sources/local/diff.go
… engine/sources/local/hash.go
… engine/sources/local/identifier.go
… engine/sources/local/localfs.go
… engine/sources/local/remotefs.go
… engine/sources/local/singleflight.go
… engine/sources/local/source.go
… go.mod
… go.sum
… sdk/elixir/lib/dagger/gen/client.ex
… sdk/elixir/lib/dagger/gen/module_source.ex
… sdk/go/dag/dag.gen.go
… sdk/go/dagger.gen.go
… sdk/php/generated/Client.php
… sdk/php/generated/ModuleSource.php