## Goals
Primary:
1. Improve caching behavior of local filesyncs; re-use cache of already synced files in more cases
2. Use the above to simplify module context/source loading + reduce unneeded cache invalidation of module builds
   * Splitting this out to a separate follow-up PR since this one is big enough on its own, but it's all setup here
3. Take control of the filesync implementation to enable future support for more features in the shell, live filesync, etc.

Secondary:
1. Improve performance of filesync in all cases (memory+CPU usage in the engine)
2. Explicitly identify and error out when concurrent changes on the client filesystem would result in inconsistent loads (previously was just undefined behavior essentially)

## Performance Improvements
I used the [git repo for the Linux kernel](https://github.com/torvalds/linux) as a stress-test case since it contains multiple GB of source files.

The main improvements are around filesyncs of already loaded data and parallel syncs, but there's some meaningful improvement to uncached syncs too for huge filesync cases.

The improvements in cache re-use and parallel syncs are expected to be most applicable to module loading and contextual directory argument usage.

* Uncached load of full repo except `.git` (`.host | directory . --exclude .git | entries`)
   * **v0.14.0** - 44.6s, ~1.3GB RSS ([trace](https://dagger.cloud/dagger/traces/b72f5fb06ff567be33d047cb10b1b77c?span=ff55914850afc762))
   * **This PR** - 29.7s, ~650MB RSS ([trace](https://dagger.cloud/dagger/traces/a56b8d4e1152eccf63295080471ffc72?span=e1b6a25d6d5ce4e4))
   * ~33% faster and 50% less memory

* Subsequent load of `drivers/` subdir (`.host | directory drivers | entries`)
  * **v0.14.0** - 20.2s ([trace](https://dagger.cloud/dagger/traces/b72f5fb06ff567be33d047cb10b1b77c?span=02c4d4c97a0446a7))
   * **This PR** - 4.9s ([trace](https://dagger.cloud/dagger/traces/a56b8d4e1152eccf63295080471ffc72?span=a70f513c9afffa70))
   * ~75% faster
   * The `drivers/` subdir was already loaded in the previous sync so there shouldn't be any loading of new data. The old implementation didn't handle this, but the new one does. You can see in the traces that the majority of the time in the new implementation is spent copying files/dirs (in order to make an immutable cache entry); this is also possible to improve in the future too.

* Cached load of full repo except `.git` using include instead of exclude (`.host | directory . --include **,!.git | entries`)
   * **v0.14.0** - 22.5s ([trace](https://dagger.cloud/dagger/traces/b72f5fb06ff567be33d047cb10b1b77c?span=dfe7df3936137b64))
   * **This PR** - 1.9s ([trace](https://dagger.cloud/dagger/traces/a56b8d4e1152eccf63295080471ffc72?span=6fe78f3345ef82f5))
   * ~91% faster
   * This case is especially silly; we are re-running the same load as before but using include settings instead of exclude settings, so there should be nothing to sync or copy. Previous implementation didn't realize this though.

* Re-run full repo load from separate parallel session (`.host | directory . --exclude .git | entries`)
   * **v0.14.0** - 21.8s ([trace](https://dagger.cloud/dagger/traces/3dc4f1dda86ea4ce8564c47699a145ea?span=64f9adb45de46174))
   * **This PR** - 1.8s ([trace](https://dagger.cloud/dagger/traces/8481833b0125ac78601d5deeac0397da?span=212c1fbbab0948c5))
   * ~91% faster
   * The previous implementation re-ran a lot of equivalent work when it happened from different clients/sessions. Concurrency across clients/sessions is handled now, which allows more work to be re-used.

## Implementation Details
Biggest problems with previous implementation:
1. Buildkit would create a new cache ref for every "name" of a provided local source
   * The name we set was derived from the path being loaded, so each path on the client's filesystem would get its own ref
   * We also couldn't use a single name and sync in different dirs under it because buildkit's cache logic requires local cache refs be locked while in use; if one is locked it just creates a brand new one (which was hard to predict and control and thus would result in lots of mysterious uncached syncs). That would have gotten much worse if we tried to maintain a single cache ref for each sync from a client.
   * Now, there's a single cache ref for each client (as identified by their stable client ID) for their entire rootfs. Each sync of an individual dir just fills the relevant part of the cache ref. Concurrency is handled allowing different syncs to safely operate on the same underlying cache ref.
   * Additionally, if the client changed files in the middle of sync previously, it was undefined what would happen. Now the implementation explicitly notices these situations and returns an error.
2. Making local syncs immutable required exporting files into the content store, which had a lot of overhead
   * We have always needed to include an extra step after a sync to copy data and make it immutable. This became our `blob` API and went through the content store for simplicity. Going through the content store involves a ton of overhead and extra copying though. It also created more situations in which the local cache sync ref was locked and would thus result in more mysterious uncached syncs.
   * Now, the copy step that creates an immutable ref is part of the filesync implementation. It uses the `contenthash` lib from upstream to calculate and store content hash metadata rather than going through the content store. It also re-uses existing equivalent refs whenever possible.

There's a ton of other little optimizations in various places (noted most of them with comments in the code).

## Future improvements and features
* Handle nested clients better?
   * Nested clients all get a brand new random stable client ID, so they don't benefit from cache reuse a ton. This is because the COW semantics of `withExec` layers complicate cache re-use quite a bit (you can end up with different filesystems that have the same stable client ID)
   * Handling this would mostly be an improvement to our integ tests (which is likely the heaviest user of nested clients by far), so may investigate this angle when looking into improving test performance
* The overhead of the copy step could be lessened by maintaining a "hardlink farm" of files stored by their content hash and just linking in the files rather than copying them on disk. 
   * In the example above, copying the 1.1GB `drivers/` subdir added ~5s of overhead, to give an idea of how much of a difference this might be able to make (not enough to be worth it here, but possibly worth it someday)
* We can enable options to force reloads of data rather than caching given ops permanently within the context of the session. This is needed for `dagger shell` use cases
* We're in more control of the filesync low-levels now, which may help "live" filesync use cases being discussed.