@helderco 2024-11-28
When doing a `dagger functions` in my local dagger/dagger clone, [I get this](https://v3.dagger.cloud/Correia/traces/265a697f6615635c7203f0dba425b5ad?listen=1b957c9883c2c799&listen=7b5d41467b47f207&listen=d5199c96e75b68cf):
```
│ │ │ ✘ upload /Users/helder/projects/dagger from hal2h27jjpkdly619pkxpnu7a (client id: 5hi0yiyaqsnmgnxfnbv4vwwzi, session id: vtnnzkwszc33xev528mgna510) (include: dagger.json, .dagger/**/*, modules/alpine/dagger.json, modules/alpine/**/*, modules/go/dagger.json, modules/go/**/*, version/dagger.json, version/**/*, modules/ps-analyzer/dagger.json, modules/ps-analyzer/**/*, sdk/python/dev/dagger.json, sdk/python/dev/**/*, cmd/dagger/dagger.json, cmd/dagger/.dagger/**/*, modules/dirdiff/dagger.json, modules/dirdiff/**/*, modules/wolfi/dagger.json, modules/wolfi/**/*, modules/golangci/dagger.json, modules/golangci/**/*, sdk/php/dev/dagger.json, sdk/php/dev/**/*, modules/shellcheck/dagger.json, modules/shellcheck/**/*, sdk/elixir/dev/dagger.json, sdk/elixir/dev/**/*, modules/graphql/dagger.json, modules/graphql/**/*, **/go.mod, **/go.sum, **/go.work, **/go.work.sum, **/vendor/, **/*.go) (exclude: **/.git, sdk/python/dev/**/.venv, sdk/python/dev/**/__pycache__, sdk/python/dev/**/.*cache) 0.0s
│ │ │ ! failed to get snapshot: failed to sync: failed to create local fs: failed to create base fs: resolve : lstat /var/lib/dagger/worker/snapshots/snapshots/83380/fs/Users: no such file or directory
│ │ │ │ ✔ filesync 0.0s
Error: failed to serve module: input: moduleSource.resolveFromCaller failed to import local module source: failed to get snapshot: failed to sync: failed to create local fs: failed to create base fs: resolve : lstat /var/lib/dagger/worker/snapshots/snapshots/83380/fs/Users: no such file or directory
```