@jedevc 2024-11-28
I do wonder how the big shared mirror of the user filesystem works with windows :thinking: Hopefully the paths all make sense.