@sipsma 2024-12-02
> When doing a dagger functions in my local dagger/dagger clone, [I get this](https://v3.dagger.cloud/Correia/traces/265a697f6615635c7203f0dba425b5ad?listen=1b957c9883c2c799&listen=7b5d41467b47f207&listen=d5199c96e75b68cf):

@helderco Thanks for trying and catching this!

Can you confirm that somewhere in the path you were loading there's a symlink? IIRC `/Users` is often a symlink on macos, but on my macbook it isn't, so not sure if I'm misremembering or what.

Either way, I created a path w/ a symlink in it, cd'd in and was able to reproduce the same error when loading a module, so pretty sure that's it. Just need to handle resolving all symlinks in the parents dirs before starting the load to avoid inconsistent behavior.