@helderco 2024-12-03
> Can you confirm that somewhere in the path you were loading there's a symlink? IIRC `/Users` is often a symlink on macos, but on my macbook it isn't, so not sure if I'm misremembering or what.

Hmm, no symlinks on macOS, but for sure in the virtual machine where the docker engine lives (OrbStack).