@sipsma 2024-12-03
> Still get the same error. No symlinks on macOS, but `/Users` is mounted in the virtual machine where the docker engine lives (OrbStack). Could that be it?

@helderco Oh interesting, glad I caught the symlink problem but must be something else left to handle...

Can you share the output of `stat /Users`? Also, how did you configure that mount of /Users into orbstack? I have orbstack too so I could try to replicate it