@sipsma 2024-12-03
> I do wonder how the big shared mirror of the user filesystem works with windows 🤔 Hopefully the paths all make sense.

Yeah I was gonna save testing Windows for a follow-up but just went ahead and did it now. Dealing with drives was not super clean unless I changed a bunch of client-side stuff in fsutil (which I've managed to avoid so far), so I ended up handling drives by just using a different filesync cache ref for each drive on a client (i.e. for client `foo` there's a separate cache ref for `C:`, `Z:`, etc.). Linux/MacOS logic stays the same and since a single filesync from a windows client is not going to span across multiple drives, I think this approach shouldn't have any downsides for windows clients.