@sipsma 2024-12-04
@helderco and I figured out what he's hitting, it has to do with case-insensitive filesystems returning inconsistent paths (thanks MacOS for that brilliant feature 🫠). Will figure out a fix and merge this after since this is definitely going to get hit by users otherwise